test: flows page for bitcoin test dapp
