+++ test/e2e/flask/btc-wallet-standard/connect.spec.ts
@@ -1,6 +1,7 @@
 // import { strict as assert } from 'assert';
 import { withBtcAccountSnap } from '../btc/common-btc';
 import { TestDappBitcoin } from '../../page-objects/pages/test-dapp-bitcoin';
+import { connectBitcoinTestDapp } from '../../page-objects/flows/bitcoin-dapp.flow';
 import { regularDelayMs } from '../../helpers';
 import { WINDOW_TITLES } from '../../constants';
 import ConnectAccountConfirmation from '../../page-objects/pages/confirmations/connect-account-confirmation';
@@ -9,7 +10,6 @@ import {
   account1Short,
   assertConnected,
   assertDisconnected,
-  connectBitcoinTestDapp,
   DEFAULT_BITCOIN_TEST_DAPP_FIXTURE_OPTIONS,
 } from './testHelpers';
