+++ test/e2e/flask/btc-wallet-standard/signMessage.spec.ts
@@ -1,11 +1,9 @@
-import { strict as assert } from 'assert';
 import { SIGNED_MESSAGES_MOCK, withBtcAccountSnap } from '../btc/common-btc';
 import { TestDappBitcoin } from '../../page-objects/pages/test-dapp-bitcoin';
+import { connectBitcoinTestDapp } from '../../page-objects/flows/bitcoin-dapp.flow';
 import { WINDOW_TITLES } from '../../constants';
 import {
-  assertConnected,
   clickConfirmButton,
-  connectBitcoinTestDapp,
   DEFAULT_BITCOIN_TEST_DAPP_FIXTURE_OPTIONS,
 } from './testHelpers';
