+++ test/e2e/flask/btc-wallet-standard/testHelpers.ts
@@ -26,85 +26,6 @@ export const DEFAULT_BITCOIN_TEST_DAPP_FIXTURE_OPTIONS = {
   },
 }
 
-/* const onboardBitcoinAccount = async (driver: Driver): Promise<void> => {
-  console.log('onboarding a new bitcoin account');
-
-  const connectAccountConfirmation = new ConnectAccountConfirmation(driver);
-  await connectAccountConfirmation.isCreateBitcoinAccountModalButtonVisible();
-  await connectAccountConfirmation.createCreateBitcoinAccountFromModal();
-}; */
-
-/* const selectAccountsAndAuthorize = async (driver: Driver): Promise<void> => {
-  console.log(
-    'select all accounts without deselecting the already selected accounts',
-  );
-  const connectAccountConfirmation = new ConnectAccountConfirmation(driver);
-  await connectAccountConfirmation.checkPageIsLoaded();
-  await connectAccountConfirmation.openEditAccountsModal();
-
-  const editConnectedAccountsModal = new EditConnectedAccountsModal(driver);
-  await editConnectedAccountsModal.checkPageIsLoaded();
-  await editConnectedAccountsModal.selectAllAccounts();
-}; */
-
-/**
- * Selects the Devnet checkbox in the permissions tab.
- *
- * @param driver
- */
-/* const selectDevnet = async (driver: Driver): Promise<void> => {
-  console.log('select devnet on permissions tab');
-
-  const connectAccountConfirmation = new ConnectAccountConfirmation(driver);
-  await connectAccountConfirmation.checkPageIsLoaded();
-  await connectAccountConfirmation.goToPermissionsTab();
-  await connectAccountConfirmation.openEditNetworksModal();
-
-  const networkPermissionSelectModal = new NetworkPermissionSelectModal(driver);
-  await networkPermissionSelectModal.checkPageIsLoaded();
-  await networkPermissionSelectModal.selectNetwork({
-    networkName: 'Bitcoin Devnet',
-  });
-  await networkPermissionSelectModal.clickConfirmEditButton();
-}; */
-
-/**
- * Connects the Bitcoin test dapp to the wallet.
- *
- * @param driver
- * @param testDapp
- * @param options
- * @param options.selectAllAccounts
- * @param options.includeDevnet
- * @param options.onboard
- */
-export const connectBitcoinTestDapp = async (
-  driver: Driver,
-  testDapp: TestDappBitcoin,
-  options: {
-    onboard?: boolean;
-    selectAllAccounts?: boolean;
-    includeDevnet?: boolean;
-    connectionLibrary?: 'sats-connect' | 'wallet-standard';
-  } = {},
-): Promise<void> => {
-  console.log('connect bitcoin test dapp');
-  await testDapp.checkPageIsLoaded();
-  await testDapp.connectToWallet(options.connectionLibrary);
-
-  // Get to extension modal, and click on the "Connect" button
-  await driver.delay(largeDelayMs);
-  await driver.switchToWindowWithTitle(WINDOW_TITLES.Dialog);
-
-  const connectAccountConfirmation = new ConnectAccountConfirmation(driver);
-  await connectAccountConfirmation.checkPageIsLoaded();
-  await connectAccountConfirmation.confirmConnect();
-
-  // Go back to the test dapp window
-  await testDapp.switchTo();
-  console.log('bitcoin test dapp connected');
-};
-
 /**
  * Waits for the Confirm button in the footer of a Bitcoin-specific modal to be clickable then clicks it.
  * Note: This function does not work for general purpose modals like connect/disconnect.
